added heatpump to services.yml
Victron is working on integrating heat pumps. I have added the “heatpump” service to services.yml.
After initialization, a heat pump device is created and data can be written. The data can be displayed in the Advanced view. It is not yet displayed on the dashboard.
Tested with version 0.9.0 and VenusOS v3.66 and v3.70.